+++ flaskapp/models/__init__.py
@@ -1,0 +1,6 @@
+# Export the things
+from .model import Model
+
+from .game import Game
+from .match import MatchTicket
+from .user import User
