+++ flaskapp/routes/api/user.py
@@ -5,181 +5,6 @@ from flaskapp.helpers import Store, Model, InvalidInputException, DuplicateExcep
 from google.cloud import datastore
 from typing import List
 
-# Define User
-class User(Model):
-    """
-    Model class for users
-    """
-
-    KIND = 'User'
-
-    def __init__(self, username="", email="", encrypted_password="", bio=""):
-        self.username = username
-        self.email = email
-        self.encrypted_password = encrypted_password
-        self.bio = bio
-        
-
-
-# Define User Storer
-class UserStore(Store):
-    """
-    Storer class for users using Google Cloud DataStore
-
-    NOTE: Inherits Store for general methods
-    """
-
-    def __init__(self):
-        pass
-
-    def post_user(self, user: User) -> User:
-        """
-        Add a user to datastore
-
-        Args:
-            user (User): User to add
-
-        Returns:
-            User: User added
-        """
-        return super().post_object(User, user)
-
-    def update_user(self, user: User) -> User:
-        """
-        Update user that exists
-
-        Args:
-            user (User): User to update
-
-        Returns:
-            User: User updated
-        """
-        return super().update_object(User, user, 'username', user.username)
-
-    def get_by_username(self, username: str) -> List[User]:
-        """
-        Get a list of users matching the username
-
-        Args:
-            username (str): Username to match
-
-        Returns:
-            List[User]: Length=1 list of users
-        """
-        return super().get_objects_by_field(User, 'username', username)
-
-    def delete_by_username(self, username: str) -> None:
-        """
-        Delete a list of users matching the username
-
-        Args:
-            username (str): Username to match
-
-        Returns:
-            List[User]: Length=1 list of users
-        """
-        return super().delete_by_field(User, 'username', username)[0]
-
-
-# Define UserService
-class UserService:
-    """
-    Service class for users.
-
-    Contains business logic and all things not directly
-    route or storage-oriented.
-    """
-
-    def __init__(self, user_store: UserStore):
-        """
-        Create a UserService
-
-        Args:
-            user_store (UserStore): Storage implementation to use
-        """
-        self.user_store = user_store
-
-    def create_user(self, user: User) -> User:
-        """
-        Create user.
-
-        Args:
-            user (User): User to create
-
-        Returns:
-            user (User): User created
-
-        Raises:
-            DuplicateException: If username exists
-            InvalidInputException: If other input issue
-            Exception: Other errors
-        """
-
-        # Make sure user's inputs exist
-        # username, epw, email
-        if not (user.username and user.email and user.encrypted_password):
-            raise InvalidInputException
-
-        # Check against db
-        if self.user_store.get_by_username(user.username):
-            raise DuplicateException
-
-        # Add user
-        return self.user_store.post_user(user)
-
-    def update_user(self, user: User) -> User:
-        """
-        Update user.
-
-        Args:
-            user (User): User to modify (username exists)
-
-        Returns:
-            user (User): User modified
-
-        Raises:
-            Exception: If user missing required field (email, username, pw),
-            or if user exists
-        """
-
-        # Make sure user's inputs exist
-        # username, epw, email
-        if not (user.username):
-            raise InvalidInputException
-
-        # Check against db
-        existing_user = self.user_store.get_by_username(user.username)[0]
-        if not existing_user:
-            raise InvalidInputException
-
-        # Update modifiable fields
-        existing_user.bio = user.bio
-        return self.user_store.update_user(existing_user)
-
-    def delete_user(self, username: str) -> User:
-        """
-        Delete user.
-
-        Args:
-            username (str): username to remove
-
-        Returns:
-            User: Removed user
-        """
-
-        # Check username doesn't exist
-        if not self.user_store.get_by_username(username):
-            raise InvalidInputException
-
-        # Delete
-        return self.user_store.delete_by_username(username)
-
-
-# SETUP ROUTES
-
-user_store = UserStore()
-user_service = UserService(user_store)
-
 # User API routes
 # POST /user : Make a user
 @app.route(API_ROOT+"/user", methods=['POST'])
