+++ flaskapp/shared.py
@@ -1,0 +1,20 @@
+from datetime import timedelta
+
+# Various exceptions for global use
+class DuplicateException(Exception):
+    pass
+
+class IllegalMoveException(Exception):
+    pass
+
+class InvalidInputException(Exception):
+    pass
+
+# NCOL, NROW
+NCOL = 7
+NROW = 6
+NCONNECT = 4
+
+# Timeouts
+GAME_TIMEOUT = timedelta(seconds = 120)
+POLLING_TIMEOUT = timedelta(seconds = 30)
