+++ flaskapp/stores/__init__.py
@@ -1,0 +1,6 @@
+# Export the things
+from .store import Store
+
+from .game import GameStore
+from .match import MatchStore
+from .user import UserStore
